face_voice_project/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
paho-mqtt==1.6.1
opencv-python==4.5.3.56
face-recognition==1.3.0
imutils==0.5.4
numpy==1.21.2
python-dotenv==0.19.0
SpeechRecognition==3.8.1
pyaudio==0.2.11